block: vhdx: Fold sector I/O into Vhdx methods
The vhdx read and write helpers lived as free functions in a separate
io.rs module, together with the Sector translation helper and the
VhdxIoError type. Free functions taking the aligned file, disk spec and
BAT as arguments are an awkward shape and the read/write names also
clash with std::io::Read and Write.

Move that logic into private read_sectors and write_sectors methods on
Vhdx, along with the Sector helper and VhdxIoError, and delete io.rs.
The read_at and write_all_at methods now call the private methods
directly. Most of the io.rs sector level tests are dropped because the
same paths are already covered by the Vhdx API tests in parser.rs. The
block allocation offset test has no equivalent, so it is ported to the
Vhdx API instead.

This is a pure refactor with no behavior change.
